requirements.txt
fastapi
jinja2
fastmvc
python-multipart
python-dotenv
google-cloud-firestore
gunicorn
uvicorn[standard]
